+++ R/imports.R
@@ -1,8 +1,5 @@
 #' @importFrom stats as.ts AIC BIC logLik nobs fitted residuals coef time
-NULL
-
 #' @importFrom zoo index
-NULL
-
+#' @importFrom xts as.xts
 #' @importFrom generics tidy augment glance
 NULL